feat(domain & data): Create Entities, and Repository
Explanation:
- Entities is using for show only data only we needed
- Repository is using for Data Source Abstraction between API and UseCase

+++ app/src/main/java/id/haaweejee/test/alfagift/alfamovie/data/source/remote/dto/DetailMovieResponse.kt
@@ -11,7 +11,7 @@ data class DetailMovieResponse(
     @SerializedName("belongs_to_collection")
     val belongsToCollection: Any? = null,
     @SerializedName("budget")
-    val budget: Int? = null,
+    val budget: Long? = null,
     @SerializedName("genres")
     val genres: List<Genre?>? = null,
     @SerializedName("homepage")
